+++ octobot_trading/personal_data/__init__.py
@@ -122,6 +122,7 @@ from octobot_trading.personal_data.portfolios import (
     BalanceProfitabilityChannel,
     SubPortfolio,
     PortfolioManager,
+    ValueConverter,
     PortfolioValueHolder,
     FuturePortfolio,
     MarginPortfolio,
@@ -328,6 +329,7 @@ __all__ = [
     "BalanceProfitabilityChannel",
     "SubPortfolio",
     "PortfolioManager",
+    "ValueConverter",
     "PortfolioValueHolder",
     "FuturePortfolio",
     "MarginPortfolio",
