+++ tests/modes/test_abstract_mode_consumer.py
@@ -136,7 +136,7 @@ async def test_get_holdings_ratio():
     exchange_manager, symbol, consumer = await _get_tools()
     exchange_manager.client_symbols = [symbol]
     exchange_manager.exchange_personal_data.portfolio_manager.portfolio_value_holder.\
-        last_prices_by_trading_pair[symbol] = decimal.Decimal("1000")
+        value_converter.last_prices_by_trading_pair[symbol] = decimal.Decimal("1000")
     exchange_manager.exchange_personal_data.portfolio_manager.portfolio_value_holder.\
         portfolio_current_value = decimal.Decimal("11")
     exchange_manager.exchange_personal_data.portfolio_manager.portfolio.portfolio = {}
